+++ CHANGELOG.md
@@ -21,5 +21,7 @@ and this project adheres to
   variables).
 - Add measurement outliers detection and removal.
 - Add standard deviation to JSON report.
+- Add `samples: { warmup: number, benchmark: number }` to run options specifying
+  the number of samples to run during warmup and benchmark.
 
 ## [0.1.11] - 2024-02-20